+++ plugins/postcss-is-pseudo-class/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changes to PostCSS Is Pseudo Class
 
+### Unreleased (patch)
+
+- Fix substitution when selectors have mixed types for compound selectors where one of two parts is a `:is()` pseudo class.
+
+```diff
+header:is(.a .b) {}
+
+/* becomes : */
+
+- header.a .b {}
++ .a header.b {}
+```
+
 ### 4.0.0
 
 _July 3, 2023_